Rename files based on riq's changes
